[ALPS05617884] network: throughput is less in DUT than REF by -90%
[Detail]
In the scenario of (USB + ipv6 only(caltd) + GRO), there will be
a lot of loss packets, cause the throughput to be very low.

The reason for the loss packet is mainly due to the clatd does
not do TCP checksum calculation for the GRO packets, causing
these GRO packages to be dropped by kernel.

[Solution]
Markup these GRO packets from clatd, and to check them at kernel.
if it is a GRO patcket,and set its ip_summed to CHECKSUM_UNNECESSARY

Change-Id: I27a9c924bd0317df49533b869a69ddb3c6a58f8c

I don't know any more information because it's a commit in the `selene-r-oss-update` branch.